Fix the growing memory problem.
A streambuf object is used to buffer outgoing socket writes. There are two streambuf objects per connection. Additionally, there is always a new connection object hanging on the
well known socket waiting for incoming connections. I was mistakenly serializing frame data to the instance that hangs on the listening socket. As a result, the internal streambuf object grew without bounds because the socket was never open.

This patch fixes the problem by not adding the new connection object to the list of active connections. Addition is delayed until the async_accept function returns control. Now, the frame data is not added to the half-initialized connection object.