Fix the half-open socket problem.
Add a timer to the TCP stream implementation that closes the socket when it fires. The initial time-out is 10 minutes. This seems reasonable for a moving craft that may intermittently lose its connection.

The timer is reset and installed whenever new data is transmitted. The timer is cancelled when the next frame is received.