SEQ frames do not include a frame trailer (END\r\n)
First, update the parsers and generators. Second, I had to change the
parse call in the TCP transport to handle multiple frames in one pass. This
should be more efficient, and it handles SEQ frames now. I still keep reading
until I encounter a frame trailer (END\r\n). The software would break if I
only looked for \r\n. However, this means that SEQ frames could be stuck in
the streambuf until a core frame is received. Practically, this is not a problem
at the moment because I ignore SEQ frames (mostly).