Implement channel closing.
Differentiate between a peer close request and a local close directive. Move
the channel objects into the channel-manager proper. Store the selected
profile in each channel so that the correct callback can be found when the
channel is closed.

Revert "Remove the profile member variable from the channel object."

This reverts commit 69e8630cb9af14849b4db702ab4c09ed16a2502b.

Conflicts:

	beep/channel.hpp